requirements.txt
flask==2.0.1
azure-kusto-data==3.1.2
azure-identity==1.7.1
python-dotenv==0.19.0
gunicorn==20.1.0